Migrate + split api_handlers into webapp/api/
