+++ webapp/anomaly_handler.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 
 	"github.com/web-platform-tests/results-analysis/metrics"
+	"github.com/web-platform-tests/wpt.fyi/webapp/api"
 	"google.golang.org/appengine"
 	"google.golang.org/appengine/datastore"
 )
@@ -22,7 +23,7 @@ type AnomalyData struct {
 // anomalyHandler handles the view of test results showing which tests pass in
 // some, but not all, browsers.
 func anomalyHandler(w http.ResponseWriter, r *http.Request) {
-	browser, err := ParseBrowserParam(r)
+	browser, err := api.ParseBrowserParam(r)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
