+++ webapp/api/github_client.go
@@ -1,4 +1,8 @@
-package webapp
+// Copyright 2017 The WPT Dashboard Project. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+package api
 
 import (
 	"fmt"
