+++ webapp/api/run_diff.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package webapp
+package api
 
 import (
 	"encoding/base64"
@@ -11,16 +11,142 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"net/url"
 	"strings"
 
 	mapset "github.com/deckarep/golang-set"
 	models "github.com/web-platform-tests/wpt.fyi/shared"
 	"golang.org/x/net/context"
+	"google.golang.org/appengine"
 	"google.golang.org/appengine/datastore"
 	"google.golang.org/appengine/urlfetch"
 )
 
-type platformAtRevision struct {
+// apiDiffHandler takes 2 test-run results JSON blobs and produces JSON in the same format, with only the differences
+// between runs.
+//
+// GET takes before and after params, for historical production runs.
+// POST takes only a before param, and the after state is provided in the body of the POST request.
+func apiDiffHandler(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case "GET":
+		handleAPIDiffGet(w, r)
+	case "POST":
+		handleAPIDiffPost(w, r)
+	default:
+		http.Error(w, fmt.Sprintf("invalid HTTP method %s", r.Method), http.StatusBadRequest)
+	}
+}
+
+func handleAPIDiffGet(w http.ResponseWriter, r *http.Request) {
+	ctx := appengine.NewContext(r)
+
+	var err error
+	params, err := url.ParseQuery(r.URL.RawQuery)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	specBefore := params.Get("before")
+	if specBefore == "" {
+		http.Error(w, "before param missing", http.StatusBadRequest)
+		return
+	}
+	var beforeJSON map[string][]int
+	if beforeJSON, err = fetchRunResultsJSONForParam(ctx, r, specBefore); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	} else if beforeJSON == nil {
+		http.Error(w, specBefore+" not found", http.StatusNotFound)
+		return
+	}
+
+	specAfter := params.Get("after")
+	if specAfter == "" {
+		http.Error(w, "after param missing", http.StatusBadRequest)
+		return
+	}
+	var afterJSON map[string][]int
+	if afterJSON, err = fetchRunResultsJSONForParam(ctx, r, specAfter); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	} else if afterJSON == nil {
+		http.Error(w, specAfter+" not found", http.StatusNotFound)
+		return
+	}
+
+	var filter DiffFilterParam
+	if filter, err = ParseDiffFilterParams(r); err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	diffJSON := getResultsDiff(beforeJSON, afterJSON, filter)
+	var bytes []byte
+	if bytes, err = json.Marshal(diffJSON); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	w.Write(bytes)
+}
+
+// handleAPIDiffPost handles POST requests to /api/diff, which allows the caller to produce the diff of an arbitrary
+// run result JSON blob against a historical production run.
+func handleAPIDiffPost(w http.ResponseWriter, r *http.Request) {
+	ctx := appengine.NewContext(r)
+
+	var err error
+	params, err := url.ParseQuery(r.URL.RawQuery)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	specBefore := params.Get("before")
+	if specBefore == "" {
+		http.Error(w, "before param missing", http.StatusBadRequest)
+		return
+	}
+	var beforeJSON map[string][]int
+	if beforeJSON, err = fetchRunResultsJSONForParam(ctx, r, specBefore); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	} else if beforeJSON == nil {
+		http.Error(w, specBefore+" not found", http.StatusNotFound)
+		return
+	}
+
+	var body []byte
+	if body, err = ioutil.ReadAll(r.Body); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	var afterJSON map[string][]int
+	if err = json.Unmarshal(body, &afterJSON); err != nil {
+		http.Error(w, "Failed to parse JSON: "+err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	var filter DiffFilterParam
+	if filter, err = ParseDiffFilterParams(r); err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	diffJSON := getResultsDiff(beforeJSON, afterJSON, filter)
+	var bytes []byte
+	if bytes, err = json.Marshal(diffJSON); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	w.Write(bytes)
+}
+
+// PlatformAtRevision is represents a test-run spec, of [platform]@[SHA],
+// e.g. 'chrome@latest' or 'safari-10@abcdef1234'
+type PlatformAtRevision struct {
 	// Platform is the string representing browser (+ version), and OS (+ version).
 	Platform string
 
@@ -28,7 +154,8 @@ type platformAtRevision struct {
 	Revision string
 }
 
-func parsePlatformAtRevisionSpec(spec string) (platformAtRevision platformAtRevision, err error) {
+// ParsePlatformAtRevisionSpec parses a test-run spec into a PlatformAtRevision struct.
+func ParsePlatformAtRevisionSpec(spec string) (platformAtRevision PlatformAtRevision, err error) {
 	pieces := strings.Split(spec, "@")
 	if len(pieces) > 2 {
 		return platformAtRevision, errors.New("invalid platform@revision spec: " + spec)
@@ -57,15 +184,15 @@ func fetchRunResultsJSONForParam(
 		}
 		return fetchRunResultsJSON(ctx, r, run)
 	}
-	var spec platformAtRevision
-	if spec, err = parsePlatformAtRevisionSpec(param); err != nil {
+	var spec PlatformAtRevision
+	if spec, err = ParsePlatformAtRevisionSpec(param); err != nil {
 		return nil, err
 	}
 	return fetchRunResultsJSONForSpec(ctx, r, spec)
 }
 
 func fetchRunResultsJSONForSpec(
-	ctx context.Context, r *http.Request, revision platformAtRevision) (results map[string][]int, err error) {
+	ctx context.Context, r *http.Request, revision PlatformAtRevision) (results map[string][]int, err error) {
 	var run models.TestRun
 	if run, err = fetchRunForSpec(ctx, revision); err != nil {
 		return nil, err
@@ -75,7 +202,7 @@ func fetchRunResultsJSONForSpec(
 	return fetchRunResultsJSON(ctx, r, run)
 }
 
-func fetchRunForSpec(ctx context.Context, revision platformAtRevision) (models.TestRun, error) {
+func fetchRunForSpec(ctx context.Context, revision PlatformAtRevision) (models.TestRun, error) {
 	baseQuery := datastore.
 		NewQuery("TestRun").
 		Order("-CreatedAt").
@@ -186,4 +313,18 @@ func anyPathMatches(paths mapset.Set, testPath string) bool {
 		}
 	}
 	return false
+}
+
+func abs(x int) int {
+	if x < 0 {
+		return -x
+	}
+	return x
+}
+
+func max(x int, y int) int {
+	if x < y {
+		return y
+	}
+	return x
 }
