+++ webapp/api/util.go
@@ -2,11 +2,13 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package webapp
+package api
 
 import (
 	"encoding/json"
 	"io/ioutil"
+	"os"
+	"path"
 	"sort"
 	"strings"
 
@@ -15,6 +17,9 @@ import (
 
 // All errors are considered fatal
 var browserNames, browserNamesAlphabetical = loadBrowsers()
+
+// ExperimentalLabel is the implicit label present for runs marked 'experimental'.
+const ExperimentalLabel = `experimental`
 
 // GetBrowserNames returns an alphabetically-ordered array of the names
 // of the browsers which are to be included (flagged as initially_loaded in the
@@ -26,8 +31,8 @@ func GetBrowserNames() ([]string, error) {
 // IsBrowserName determines whether the given name string is a valid browser name.
 // Used for validating user-input params for browsers.
 func IsBrowserName(name string) bool {
-	if strings.HasSuffix(name, "-"+experimentalLabel) {
-		name = name[0 : len(name)-1-len(experimentalLabel)] // Trim suffix.
+	if strings.HasSuffix(name, "-"+ExperimentalLabel) {
+		name = name[0 : len(name)-1-len(ExperimentalLabel)] // Trim suffix.
 	}
 	_, ok := browserNames[name]
 	return ok
@@ -42,7 +47,15 @@ func loadBrowsers() (map[string]bool, []string) {
 	var browserNames map[string]bool
 	var browserNamesAlphabetical []string
 
-	if bytes, err = ioutil.ReadFile("browsers.json"); err != nil {
+	// Climb the working directory into /webapp/ folder
+	var dir string
+	if dir, _ = os.Getwd(); err != nil {
+		panic(err)
+	}
+	for path.Base(dir) != "webapp" {
+		dir = path.Join(dir, "..")
+	}
+	if bytes, err = ioutil.ReadFile(path.Join(dir, "browsers.json")); err != nil {
 		panic(err)
 	}
 
@@ -60,17 +73,4 @@ func loadBrowsers() (map[string]bool, []string) {
 	sort.Strings(browserNamesAlphabetical)
 
 	return browserNames, browserNamesAlphabetical
-}
-
-func abs(x int) int {
-	if x < 0 {
-		return -x
-	}
-	return x
-}
-func max(x int, y int) int {
-	if x < y {
-		return y
-	}
-	return x
 }
